Rehydrate a ggseg atlas into sf-backed form

Inverse of as_polygon_atlas(). Sets the single geom slot to an sf-class geometry table, converting from polygons via sfheaders::sf_multipolygon() if needed — sfheaders is pure Rcpp with no GDAL/GEOS/PROJ dependencies, so the conversion itself does not require a full sf installation. Use this when you want to run sf operations (buffers, intersections, CRS transforms) on atlas geometry; those sf operations themselves still require sf.

Usage

as_sf_atlas(atlas)

Arguments

atlas

A ggseg_atlas (or legacy brain_atlas) object.

Value

A ggseg_atlas whose $data$geom is an sf object.

Details

Conversion is lossless, so a single representation is kept (no redundant polygons alongside sf).

Examples

atlas <- as_sf_atlas(as_polygon_atlas(dk()))
sf::st_buffer(atlas_geom(atlas)$geometry[[1]], dist = 2)
#> POLYGON ((59.43094 36.12883, 59.32851 36.15492, 59.22759 36.18638, 59.12848 36.22313, 59.03145 36.26506, 58.93677 36.31206, 58.8447 36.364, 58.75551 36.42073, 58.66944 36.48209, 58.58673 36.54792, 58.50762 36.61802, 58.43231 36.69221, 58.36104 36.77027, 58.29399 36.85198, 58.23135 36.93713, 58.17329 37.02546, 58.11998 37.11674, 58.07157 37.21071, 58.0282 37.30711, 57.98997 37.40566, 57.28891 39.36531, 57.25635 39.46438, 57.22901 39.565, 57.20695 39.66692, 57.19022 39.76984, 57.17889 39.8735, 57.17297 39.97761, 57.17249 40.08188, 57.17744 40.18604, 57.18782 40.28979, 57.20358 40.39287, 57.2247 40.49498, 57.25112 40.59586, 57.28275 40.69522, 57.31952 40.7928, 57.36133 40.88832, 57.40805 40.98154, 57.45958 41.0722, 57.51575 41.16005, 57.57643 41.24485, 57.64145 41.32637, 57.71063 41.4044, 57.78378 41.47871, 57.8607 41.54911, 57.94119 41.6154, 59.5424 42.8656, 57.00199 47.5496, 56.95117 47.64994, 56.90607 47.75298, 56.86683 47.85839, 56.83358 47.96584, 56.80642 48.07499, 56.78544 48.18549, 56.63805 49.10185, 56.62412 49.20571, 56.61566 49.31016, 56.61267 49.4149, 56.61518 49.51966, 56.62317 49.62415, 56.63662 49.72807, 56.6555 49.83115, 56.67975 49.93309, 56.7093 50.03363, 56.74408 50.13248, 56.78399 50.22937, 56.82892 50.32404, 56.87875 50.41622, 56.93334 50.50567, 56.99254 50.59214, 57.05618 50.67539, 57.1241 50.75519, 57.1961 50.83132, 57.272 50.90358, 57.35157 50.97176, 57.4346 51.03568, 57.52087 51.09517, 57.61014 51.15005, 57.70216 51.20019, 57.79668 51.24543, 57.89344 51.28566, 59.01609 51.71833, 58.51508 53.25104, 58.48557 53.34987, 58.4612 53.45008, 58.44202 53.55142, 58.4281 53.65361, 58.41946 53.75638, 58.41613 53.85946, 58.41812 53.96257, 58.42542 54.06545, 58.43802 54.16781, 58.45587 54.26939, 58.47894 54.36991, 58.50716 54.46911, 58.54046 54.56672, 58.57874 54.66248, 58.62191 54.75615, 58.66985 54.84746, 58.72243 54.93619, 58.77952 55.02208, 58.84096 55.10492, 58.90659 55.18447, 58.97623 55.26054, 59.0497 55.33292, 59.12681 55.40142, 59.20734 55.46585, 59.29108 55.52604, 59.37782 55.58184, 59.46732 55.63309, 59.55934 55.67965, 59.65364 55.72142, 59.74997 55.75826, 60.24609 55.9335, 59.79535 57.06848, 59.759 57.16754, 59.72793 57.26838, 59.70222 57.37071, 59.68195 57.47426, 59.66716 57.57874, 59.6579 57.68385, 59.6542 57.7893, 59.65607 57.8948, 59.6635 58.00006, 59.67647 58.10477, 59.69494 58.20866, 59.71887 58.31143, 59.74819 58.41279, 59.78281 58.51247, 59.82263 58.61018, 59.86756 58.70566, 59.91746 58.79863, 59.97219 58.88884, 60.03161 58.97604, 60.09554 59.05998, 60.16381 59.14044, 60.23622 59.21718, 60.31259 59.29, 60.39269 59.35869, 60.4763 59.42306, 60.56318 59.48293, 60.65311 59.53813, 60.74582 59.58851, 61.913 60.18349, 61.82464 62.35844, 61.82311 62.46139, 61.82688 62.56429, 61.83594 62.66686, 61.85027 62.76882, 61.86982 62.86992, 61.89456 62.96987, 61.9244 63.06841, 61.95928 63.16529, 61.9991 63.26025, 62.04375 63.35303, 62.09311 63.44339, 62.14707 63.53109, 62.20546 63.6159, 62.26814 63.69759, 62.33495 63.77594, 62.40569 63.85075, 62.4802 63.92183, 62.55826 63.98897, 62.63967 64.05201, 62.72422 64.11077, 62.81169 64.16511, 62.90183 64.21487, 65.51519 65.57092, 65.6076 65.61589, 65.70219 65.65606, 65.79872 65.69132, 65.89694 65.72157, 65.99658 65.74673, 66.09738 65.76675, 66.19908 65.78156, 66.30141 65.79112, 66.40409 65.79542, 66.50685 65.79444, 66.60943 65.78817, 66.71155 65.77665, 66.81295 65.7599, 66.91335 65.73796, 67.01249 65.7109, 67.11011 65.67877, 67.20595 65.64168, 67.29976 65.5997, 67.39129 65.55297, 67.56301 65.45966, 68.68108 66.44719, 68.76271 66.51551, 68.84786 66.57939, 68.93629 66.63865, 69.02774 66.69312, 69.12197 66.74265, 69.21869 66.78709, 69.31764 66.82633, 69.41855 66.86024, 69.52111 66.88874, 69.62504 66.91174, 69.73005 66.92918, 69.83584 66.941, 69.94211 66.94719, 70.04855 66.9477, 70.15488 66.94256, 70.26077 66.93176, 70.36595 66.91535, 70.4701 66.89336, 70.57294 66.86587, 70.67416 66.83294, 70.77349 66.79467, 70.87065 66.75117, 70.96535 66.70256, 71.05733 66.64899, 71.14633 66.59059, 73.32117 65.07937, 73.40383 65.01874, 73.48327 64.95395, 73.55928 64.88518, 73.63167 64.8126, 73.70025 64.73641, 73.76483 64.65681, 73.82525 64.574, 73.88135 64.4882, 73.93298 64.39964, 76.73895 59.28993, 80.21285 58.13473, 80.31738 58.09671, 80.41963 58.05294, 80.51928 58.00356, 80.61604 57.94871, 80.7096 57.88856, 80.79967 57.82331, 80.88598 57.75316, 80.96825 57.67831, 81.69897 56.97542, 85.15105 55.91886, 89.22209 55.2903, 98.5048 55.26443, 107.6116 56.45479, 107.6618 56.4607, 113.6546 57.09058, 113.8019 57.10058, 123.8462 57.41101, 123.9364 57.41177, 124.0265 57.40845, 124.1163 57.40108, 128.7052 56.92034, 128.8048 56.90738, 128.9036 56.88943, 133.4384 55.94747, 133.5481 55.92143, 133.6562 55.88926, 133.7623 55.85104, 140.0654 53.37745, 140.166 53.33475, 140.2642 53.28661, 140.3595 53.23318, 144.6683 50.66224, 144.7575 50.60573, 144.8437 50.54458, 144.9265 50.47897, 145.0057 50.40907, 145.0812 50.33509, 145.1526 50.25723, 145.2198 50.1757, 145.2826 50.09074, 145.3408 50.00259, 145.3943 49.91148, 145.4429 49.81767, 145.4865 49.72143, 145.525 49.62302, 145.5582 49.52271, 145.586 49.42079, 145.6084 49.31755, 145.6254 49.21326, 145.6368 49.10823, 145.6426 49.00274, 145.6429 48.89709, 145.6376 48.79157, 145.6267 48.68647, 145.6103 48.5821, 145.5884 48.47874, 145.5686 48.40467, 146.3163 47.95183, 146.4039 47.89563, 146.4884 47.83496, 146.5697 47.76997, 146.6474 47.70084, 146.7215 47.62776, 146.7917 47.55093, 146.8577 47.47056, 146.9196 47.38685, 146.9769 47.30005, 147.0297 47.21038, 147.0777 47.11809, 147.1209 47.02342, 147.1591 46.92663, 147.1922 46.82799, 147.2202 46.72777, 147.2428 46.62622, 147.2602 46.52363, 147.2722 46.42028, 147.2789 46.31644, 147.2801 46.2124, 147.2759 46.10844, 147.2663 46.00483, 147.2513 45.90187, 147.231 45.79982, 147.2054 45.69897, 147.1746 45.59958, 147.1387 45.50193, 147.0977 45.40628, 147.0701 45.35014, 147.0865 45.34108, 147.1752 45.28579, 147.2609 45.22593, 147.3434 45.16168, 147.4224 45.09321, 147.4977 45.0207, 147.5691 44.94436, 147.6364 44.8644, 147.6995 44.78102, 147.7581 44.69447, 147.8121 44.60497, 147.8613 44.51277, 147.9057 44.41812, 147.9451 44.32128, 147.9793 44.22252, 148.0084 44.12211, 148.0321 44.02031, 148.0505 43.91741, 148.0636 43.81369, 148.0711 43.70943, 148.0732 43.60492, 148.0699 43.50044, 148.0611 43.39628, 148.0469 43.29272, 148.0272 43.19005, 148.0023 43.08854, 147.972 42.98847, 147.9366 42.89012, 147.8961 42.79376, 147.8506 42.69964, 147.8003 42.60803, 147.7452 42.51917, 147.6856 42.43332, 147.6215 42.35069, 147.5533 42.27153, 147.481 42.19604, 147.4048 42.12443, 147.325 42.05689, 147.2418 41.99362, 147.1554 41.93479, 147.0661 41.88054, 146.974 41.83104, 146.8795 41.78642, 146.7827 41.7468, 146.6841 41.71228, 146.5837 41.68297, 139.5957 39.8384, 139.5152 39.81893, 120.1403 35.55385, 120.0384 35.53419, 119.9357 35.51983, 119.8324 35.51082, 119.7287 35.50717, 119.625 35.50891, 119.5216 35.51601, 119.4186 35.52847, 119.3164 35.54625, 119.2153 35.56931, 119.1155 35.59757, 119.0625 35.61562, 106.7516 32.39095, 106.658 32.36883, 106.5635 32.35123, 103.2212 31.8118, 87.63629 28.10097, 87.53511 28.07963, 87.43297 28.06354, 87.33013 28.05276, 87.22687 28.0473, 87.12346 28.04719, 87.02019 28.05243, 86.91733 28.06299, 86.81515 28.07886, 86.71393 28.09999, 86.61393 28.12632, 86.55767 28.14429, 79.23229 26.56171, 79.12386 26.5414, 79.01448 26.5271, 67.27535 25.32031, 67.18135 25.31289, 67.0871 25.3099, 66.99281 25.31135, 56.90878 25.70501, 56.80484 25.71178, 56.70139 25.72396, 56.59872 25.7415, 56.4971 25.76436, 56.39681 25.79249, 56.29812 25.8258, 56.20129 25.8642, 56.1066 25.90759, 56.0143 25.95585, 55.92463 26.00885, 55.83784 26.06645, 55.75417 26.12848, 55.67384 26.19479, 55.59708 26.2652, 55.52408 26.3395, 55.45505 26.4175, 55.39018 26.49899, 55.32963 26.58375, 55.27358 26.67155, 55.22218 26.76214, 55.17557 26.85529, 55.13386 26.95074, 55.09719 27.04822, 55.06563 27.14749, 55.03929 27.24827, 55.01823 27.35028, 55.00251 27.45324, 54.99218 27.55689, 54.98725 27.66093, 54.98775 27.76509, 54.99366 27.86909, 55.00499 27.97263, 55.02169 28.07544, 55.04372 28.17725, 55.07102 28.27777, 55.10352 28.37673, 55.14112 28.47386, 55.18373 28.56891, 55.23123 28.66161, 55.2835 28.75171, 55.34038 28.83897, 55.40173 28.92314, 55.46738 29.00401, 55.53715 29.08135, 55.61085 29.15496, 55.68828 29.22462, 55.76924 29.29017, 55.8535 29.3514, 55.94083 29.40817, 56.031 29.46031, 56.12376 29.50769, 56.21886 29.55018, 56.31605 29.58766, 59.7121 30.79663, 59.71763 30.86504, 59.73152 30.96882, 59.75081 31.07173, 59.77547 31.17349, 59.80541 31.27382, 59.84057 31.37244, 59.88084 31.46909, 59.92611 31.5635, 59.97626 31.65541, 60.03116 31.74457, 60.09064 31.83074, 60.15455 31.91367, 60.22271 31.99315, 60.29494 32.06895, 60.37104 32.14087, 60.4508 32.2087, 60.53399 32.27227, 60.6204 32.3314, 60.70979 32.38592, 60.80191 32.43569, 60.8965 32.48057, 60.99332 32.52044, 62.83793 33.2242, 62.76976 33.27952, 62.69908 33.34267, 59.8196 36.04073, 59.43094 36.12883))